remove instructions for adding a github webhook secret
there is no need to do the step indicated in exercise 1. the secret
associated with openshift's URL and github's secret are not the same.

From the docs at: https://docs.openshift.com/container-platform/3.11/dev_guide/builds/triggering_builds.html#github-webhooks

"The secret used in the webhook trigger configuration is not the same
as secret field you encounter when configuring webhook in GitHub UI.
The former is to make the webhook URL unique and hard to predict,
the latter is an optional string field used to create HMAC hex digest of the body,
which is sent as an X-Hub-Signature header."

Signed-off-by: Steve Martinelli <s.martinelli@gmail.com>